requirements.txt
numpy==1.24.1
pandas==1.5.2
scikit-learn==1.2.2
hydra-core==1.3.2
omegaconf==2.3.0
mlflow==2.3.2
sqlalchemy==1.4.46
psycopg2-binary==2.9.6
pylint==2.15.0
pytest-cov==4.0.0
tqdm==4.65.0
geopy==2.3.0
requests==2.30.0
xgboost==1.7.5
streamlit==1.22.0
shap==0.41.0
interpret==0.3.0
hydra-optuna-sweeper==1.2.0
uvicorn==0.22.0
fastapi==0.95.2
